2.06.32415.1345
Fixed Demon Invasion not starting on its own (because of an option that
was removed recently -- older users still have the setting in their
Windows registry and new users don't, so it's working fine for most
older users).  Also extended individual deck support to Lies of
Astaroth.  Shiva->Destroyer demon name change supported internally (was
one spot still referencing the old name which would be caught if the
Destroyer demon was in progress during login).