Tweaks to enchanting
Allow enchanting to succeed if the server preview level is less than the
requested enchantment level, as long as it's higher than the card's
current level.  This is a server issue we're trying to work around which
may require users to enchant the same card more than once (rather than
continually failing on certain cards).